+++ apps/transfer/src/framework-config.ts
@@ -1,0 +1,29 @@
+import {
+  ComponentRenderArgs,
+  IAppConfigurator,
+} from '@equinor/fusion-framework-react-app';
+import { enableContext } from '@equinor/fusion-framework-react-module-context';
+import buildQuery from 'odata-query';
+
+export const configure = async (config: IAppConfigurator, c: ComponentRenderArgs) => {
+  enableContext(config, async (builder) => {
+    builder.setContextType(['ProjectMaster', 'Facility']);
+    builder.setContextParameterFn(({ search, type }) => {
+      return buildQuery({
+        search,
+        filter: {
+          type: {
+            in: type,
+          },
+        },
+      });
+    });
+  });
+
+  config.configureHttpClient('cc-api', {
+    baseUri: "https://famapi.equinor.com",
+    defaultScopes: ["api://958bef40-48c3-496e-bc0b-0fe5783f196b/access_as_user"],
+  });
+
+};
+
